fix(Updater): Handle session death
Fixes #3091

+++ app/src/main/java/app/revanced/manager/ui/viewmodel/UpdateViewModel.kt
@@ -2,12 +2,14 @@ package app.revanced.manager.ui.viewmodel
 
 import android.app.Application
 import android.net.Uri
+import android.os.ParcelUuid
 import androidx.annotation.StringRes
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableLongStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
+import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.Pager
@@ -35,6 +37,7 @@ import kotlinx.coroutines.withContext
 import ru.solrudev.ackpine.installer.InstallFailure
 import ru.solrudev.ackpine.installer.PackageInstaller
 import ru.solrudev.ackpine.installer.createSession
+import ru.solrudev.ackpine.installer.getSession
 import ru.solrudev.ackpine.session.Session
 import ru.solrudev.ackpine.session.await
 import ru.solrudev.ackpine.session.parameters.Confirmation
@@ -49,22 +52,24 @@ class UpdateViewModel(
     private val fs: Filesystem,
     private val ackpineInstaller: PackageInstaller,
     private val managerUpdateRepository: ManagerUpdateRepository,
+    private val savedStateHandle: SavedStateHandle,
 ) : ViewModel() {
-    // TODO: save state to handle process death.
     var downloadedSize by mutableLongStateOf(0L)
         private set
     var totalSize by mutableLongStateOf(0L)
         private set
+
     val downloadProgress by derivedStateOf {
         if (downloadedSize == 0L || totalSize == 0L) return@derivedStateOf 0f
-
         downloadedSize.toFloat() / totalSize.toFloat()
     }
+
     var showInternetCheckDialog by mutableStateOf(false)
     var state by mutableStateOf(State.CAN_DOWNLOAD)
         private set
 
     var installError by mutableStateOf("")
+        private set
 
     var releaseInfo: ReVancedAsset? by mutableStateOf(null)
         private set
@@ -79,7 +84,37 @@ class UpdateViewModel(
 
     private val location = fs.tempDir.resolve("updater.apk")
 
+    private var installerSessionId: ParcelUuid? = savedStateHandle["installerSessionId"]
+        set(value) {
+            field = value
+            savedStateHandle["installerSessionId"] = value
+        }
+
     init {
+        installerSessionId?.uuid?.let { id ->
+            viewModelScope.launch {
+                uiSafe(app, R.string.install_app_fail, "Failed to install") {
+                    state = State.INSTALLING
+
+                    val session = withContext(Dispatchers.IO) {
+                        ackpineInstaller.getSession(id)
+                    }
+
+                    if (session == null) {
+                        installerSessionId = null
+                        state = State.CAN_INSTALL
+                        return@uiSafe
+                    }
+
+                    val result = withContext(Dispatchers.IO) {
+                        session.await()
+                    }
+
+                    handleInstallResult(result)
+                }
+            }
+        }
+
         viewModelScope.launch {
             uiSafe(app, R.string.download_manager_failed, "Failed to download ReVanced Manager") {
                 releaseInfo = managerUpdateRepository.getUpdateOrNull()
@@ -87,6 +122,8 @@ class UpdateViewModel(
 
                 if (downloadOnScreenEntry) {
                     downloadUpdate()
+                } else if (location.exists()) {
+                    state = State.CAN_INSTALL
                 } else {
                     state = State.CAN_DOWNLOAD
                 }
@@ -97,43 +134,63 @@ class UpdateViewModel(
     fun downloadUpdate(ignoreInternetCheck: Boolean = false) = viewModelScope.launch {
         uiSafe(app, R.string.failed_to_download_update, "Failed to download update") {
             val release = releaseInfo!!
-            withContext(Dispatchers.IO) {
-                if (!ignoreInternetCheck && !networkInfo.isUnmetered()) {
-                    showInternetCheckDialog = true
-                } else {
-                    state = State.DOWNLOADING
 
-                    http.download(location) {
-                        url(release.downloadUrl)
-                        onDownload { bytesSentTotal, contentLength ->
-                            withContext(Dispatchers.Main) {
-                                downloadedSize = bytesSentTotal
-                                contentLength?.let { totalSize = it }
-                            }
-                        }
+            if (!ignoreInternetCheck && !networkInfo.isUnmetered()) {
+                showInternetCheckDialog = true
+                return@uiSafe
+            }
+
+            state = State.DOWNLOADING
+
+            withContext(Dispatchers.IO) {
+                http.download(location) {
+                    url(release.downloadUrl)
+                    onDownload { bytesSentTotal, contentLength ->
+                        downloadedSize = bytesSentTotal
+                        contentLength?.let { totalSize = it }
                     }
-                    installUpdate()
                 }
             }
+
+            installUpdate()
         }
     }
 
     fun installUpdate() = viewModelScope.launch {
-        state = State.INSTALLING
-        val result = withContext(Dispatchers.IO) {
-            ackpineInstaller.createSession(Uri.fromFile(location)) {
-                confirmation = Confirmation.IMMEDIATE
-            }.await()
-        }
+        uiSafe(app, R.string.install_app_fail, "Failed to install") {
+            state = State.INSTALLING
 
+            val session = withContext(Dispatchers.IO) {
+                ackpineInstaller.createSession(Uri.fromFile(location)) {
+                    confirmation = Confirmation.IMMEDIATE
+                }
+            }
+
+            installerSessionId = ParcelUuid(session.id)
+
+            val result = try {
+                withContext(Dispatchers.IO) {
+                    session.await()
+                }
+            } finally {
+                installerSessionId = null
+            }
+
+            handleInstallResult(result)
+        }
+    }
+
+    private fun handleInstallResult(result: Session.State<InstallFailure>) {
         when (result) {
-            is Session.State.Failed<InstallFailure> -> when (val failure = result.failure) {
-                is InstallFailure.Aborted -> state = State.CAN_INSTALL
-                else -> {
-                    val msg = failure.message.orEmpty()
-                    app.toast(app.getString(R.string.install_app_fail, msg))
-                    installError = msg
-                    state = State.FAILED
+            is Session.State.Failed<InstallFailure> -> {
+                when (val failure = result.failure) {
+                    is InstallFailure.Aborted -> state = State.CAN_INSTALL
+                    else -> {
+                        val msg = failure.message.orEmpty()
+                        app.toast(app.getString(R.string.install_app_fail, msg))
+                        installError = msg
+                        state = State.FAILED
+                    }
                 }
             }
 
@@ -141,12 +198,18 @@ class UpdateViewModel(
                 app.toast(app.getString(R.string.install_app_success))
                 state = State.SUCCESS
             }
+
+            else -> {
+                state = State.INSTALLING
+            }
         }
     }
 
     override fun onCleared() {
         super.onCleared()
-        location.delete()
+        if (installerSessionId == null) {
+            location.delete()
+        }
     }
 
     enum class State(@param:StringRes val title: Int) {
